requirements.txt
alembic==1.6.5
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Markdown==0.3
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-PageDown==0.4.0
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-editor==1.0.4
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
